move localtest renderer into ageneric resourcegraph renderer
