+++ src/cli/internal/ui/resourcegraph/compact.go
@@ -1,4 +1,4 @@
-package renderer
+package resourcegraph
 
 import (
 	"fmt"
