+++ src/cli/internal/ui/resourcegraph/log.go
@@ -1,4 +1,4 @@
-package renderer
+package resourcegraph
 
 import (
 	"strconv"
