+++ src/cli/internal/ui/resourcegraph/model.go
@@ -1,4 +1,4 @@
-package renderer
+package resourcegraph
 
 import (
 	"strings"
